refactor(protocol): all chain ID is now uint64 (#15079)

+++ packages/protocol/genesis/GenerateGenesis.g.sol
@@ -372,7 +372,7 @@ contract TestGenerateGenesis is Test, AddressResolver {
     {
         assertEq(
             getPredeployedContractAddress(contractName),
-            addressManager.getAddress(block.chainid, name)
+            addressManager.getAddress(uint64(block.chainid), name)
         );
     }
 }
